+++ tests/TestCase.php
@@ -6,6 +6,9 @@ use Lunar\Facades\ModelManifest;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 use Spatie\StructureDiscoverer\Discover;
 
+use function Orchestra\Testbench\after_resolving;
+use function Orchestra\Testbench\default_migration_path;
+
 class TestCase extends BaseTestCase
 {
     protected function setUp(): void
@@ -16,6 +19,25 @@ class TestCase extends BaseTestCase
     protected function getEnvironmentSetUp($app)
     {
         $this->replaceModelsForTesting();
+
+        // File-backed SQLite per worker; Testbench wipes RefreshDatabase's
+        // cache for `:memory:`, forcing a full migrate every test.
+        $dbPath = sys_get_temp_dir().'/lunar-test-'.getmypid().'.sqlite';
+        if (! file_exists($dbPath)) {
+            touch($dbPath);
+        }
+
+        $app['config']->set('database.connections.testing.database', $dbPath);
+    }
+
+    // Register laravel migrations on the migrator instead of running them
+    // separately — a standalone migrate commits DDL and resets RefreshDatabase's
+    // per-process cache.
+    protected function defineDatabaseMigrations(): void
+    {
+        after_resolving($this->app, 'migrator', static function ($migrator) {
+            $migrator->path(default_migration_path());
+        });
     }
 
     /**
